convert to flaot for explainer

+++ dianna/dashboard/pages/Tabular.py
@@ -93,7 +93,7 @@ grid_response = AgGrid(
 
 if grid_response['selected_rows'] is not None:
     selected_row = grid_response['selected_rows']['Index'].iloc[0]
-    selected_data = data.iloc[selected_row, 1:].to_numpy()
+    selected_data = data.iloc[selected_row, 1:].to_numpy(dtype=np.float32)
     with st.spinner('Predicting class'):
         predictions = predict(model=serialized_model, tabular_input=selected_data)
